+++ src/main/java/org/boutwaretech/weightspec/repositories/BodyMeasurementTransactionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "transactions", path = "transactions")
-public interface BodyMeasurementTransactionRepository extends PagingAndSortingRepository<BodyMeasurementTransaction, Integer> {
+public interface BodyMeasurementTransactionRepository extends PagingAndSortingRepository<BodyMeasurementTransaction, Long> {
 
 }
